requirements.txt
Flask==0.10.1
Flask-WTF==0.9.4
Jinja2==2.7.2
MarkupSafe==0.18
Pillow==2.3.0
WTForms==1.0.5
Werkzeug==0.9.4
itsdangerous==0.23
wsgiref==0.1.2
